Enabled Close+Sync for improved persistence & temp fix for failure handling.
